SystemUI: migrate statusbar manual brightness to float
Google switched to float in BrightnessController after android11,
so stay inline with that behavior. Besides, the brightness slider
is in log scale since pie, and a gamma-to-linear conversion is
performed when setting the value to DisplayManager. Therefore, use
the newly introduced float conversion here to achieve a perception
of linear brightness transitions to human eyes.

Change-Id: I70d5e33b9464c73a25f73a1d9fc690c101b1fa05
Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>